requirements.txt
amqp==2.2.2
Babel==2.5.1
billiard==3.5.0.3
celery==4.1.0
certifi==2017.11.5
chardet==3.0.4
Django==1.11.7
django-autoslug==1.9.3
django-taggit==0.22.1
idna==2.6
kombu==4.1.0
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
six==1.11.0
Unidecode==0.4.21
urllib3==1.22
vine==1.1.4
